+++ backfit_generator/example/example.dart
@@ -1,7 +1,9 @@
 
+
 import 'dart:io';
 
 import 'package:backfit/backfit.dart';
+import 'package:cross_file/cross_file.dart';
 
 import 'my_post.dart';
 
@@ -10,9 +12,9 @@ part 'example.backfit.dart';
 @BackfitService()
 abstract class Posts {
   @Get('posts')
-  Future<Response<MyPost>> postsByUserId(@Query('userId') String userId); 
+  Future<Response<MyPost>> postsByUserId(@Query('userId') String userId, @Header('content-type') String contentType); 
 
   @multiPart
   @Post('photos')
-  Future<Response> uploadFile(@PartFile("image", "media/*") File file);
+  Future<Response> uploadFile(@PartFile("image", "media/*") XFile? file);
 }