+++ example/lib/posts_service.backfit.dart
@@ -19,13 +19,21 @@ class PostsService extends Posts with _$Posts {
 mixin _$Posts implements Posts {
   BackfitClient? _client;
 
+  final List<int> _validStatuses = [
+    HttpStatus.ok,
+    HttpStatus.created,
+    HttpStatus.accepted,
+    HttpStatus.nonAuthoritativeInformation,
+    HttpStatus.partialContent,
+  ];
+
   @override
   Future<Response<List<MyPost>>> postsByUserId(String userId) async {
     final res = await _client!
         .get(Uri.parse('${_client!.baseUrl}/posts?userId=$userId'));
 
     return Response(
-      data: res.body.isNotEmpty
+      data: _validStatuses.contains(res.statusCode) && res.body.isNotEmpty
           ? (json.decode(res.body) as List)
               .map((it) => MyPost.fromJson(it))
               .toList()
@@ -46,7 +54,9 @@ mixin _$Posts implements Posts {
     final data = await res.stream.bytesToString();
 
     return Response(
-      data: data.isNotEmpty ? json.decode(data) : null,
+      data: _validStatuses.contains(res.statusCode) && data.isNotEmpty
+          ? json.decode(data)
+          : null,
       statusCode: res.statusCode,
       reasonPhrase: res.reasonPhrase,
       headers: HttpHeaders.fromMap(res.headers),
@@ -65,7 +75,23 @@ mixin _$Posts implements Posts {
     final data = await res.stream.bytesToString();
 
     return Response(
-      data: data.isNotEmpty ? json.decode(data) : null,
+      data: _validStatuses.contains(res.statusCode) && data.isNotEmpty
+          ? json.decode(data)
+          : null,
+      statusCode: res.statusCode,
+      reasonPhrase: res.reasonPhrase,
+      headers: HttpHeaders.fromMap(res.headers),
+    );
+  }
+
+  @override
+  Future<Response> addPost(MyPost post) async {
+    final res = await _client!
+        .post(Uri.parse('${_client!.baseUrl}/posts'), body: post.toJson());
+    return Response(
+      data: _validStatuses.contains(res.statusCode) && res.body.isNotEmpty
+          ? json.decode(res.body)
+          : null,
       statusCode: res.statusCode,
       reasonPhrase: res.reasonPhrase,
       headers: HttpHeaders.fromMap(res.headers),