+++ backfit/CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.0.2+8] 15/06/2022
+
+* Added support for Header param.
+## [0.0.2+7] 07/04/2022
+
+* Fixed PartField annotation.
+## [0.0.2+6] 29/03/2022
+
+* Updated to use the latest analyzer
+## [0.0.2+5] 09/12/2021
+
+* Hotfix for Multipart Requests in Pretty Logger
+
+## [0.0.2+4] 02/12/2021
+
+* Added request logging in Pretty Logger
 ## [0.0.2+3] 17/09/2021
 
 * Multipart headers support.